Moving bars into composables.
