+++ app/src/main/java/com/adammcneilly/pocketleague/ui/OnePane.kt
@@ -1,11 +1,13 @@
 package com.adammcneilly.pocketleague.ui
 
+import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.saveable.SaveableStateHolder
 import androidx.compose.ui.Modifier
+import com.adammcneilly.pocketleague.composables.Level1BottomBar
 import com.adammcneilly.pocketleague.composables.TopBar
 import com.adammcneilly.pocketleague.shared.screens.Navigation
 
@@ -35,8 +37,13 @@ fun Navigation.OnePane(
         },
         bottomBar = {
             if (currentScreenIdentifier.screen.navigationLevel == 1) {
-                Level1BottomBar(currentScreenIdentifier)
+                Level1BottomBar(
+                    selectedTab = currentScreenIdentifier,
+                    modifier = Modifier
+                        .navigationBarsPadding(),
+                )
             }
         },
+        modifier = modifier,
     )
 }
