+++ composables/src/commonMain/kotlin/com/adammcneilly/pocketleague/composables/Level1BottomBar.kt
@@ -1,11 +1,9 @@
-package com.adammcneilly.pocketleague.ui
+package com.adammcneilly.pocketleague.composables
 
-import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.BarChart
 import androidx.compose.material.icons.filled.Feed
 import androidx.compose.material.icons.filled.Leaderboard
-import androidx.compose.material3.BottomAppBar
 import androidx.compose.material3.Icon
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
@@ -16,12 +14,14 @@ import com.adammcneilly.pocketleague.shared.screens.Level1Navigation
 import com.adammcneilly.pocketleague.shared.screens.Navigation
 import com.adammcneilly.pocketleague.shared.screens.ScreenIdentifier
 
+
 /**
- * Creates a [BottomAppBar] to manage the menu items on a phone or other compact screen size.
+ * Creates a [NavigationBar] to manage the menu items on a phone or other compact screen size.
  */
 @Composable
 fun Navigation.Level1BottomBar(
-    selectedTab: ScreenIdentifier
+    selectedTab: ScreenIdentifier,
+    modifier: Modifier,
 ) {
     NavigationBar(
         content = {
@@ -44,7 +44,6 @@ fun Navigation.Level1BottomBar(
                 onClick = { navigateByLevel1Menu(Level1Navigation.Records) },
             )
         },
-        modifier = Modifier
-            .navigationBarsPadding(),
+        modifier = modifier,
     )
 }
