+++ composables/src/commonMain/kotlin/com/adammcneilly/pocketleague/composables/Level1NavigationRail.kt
@@ -1,4 +1,4 @@
-package com.adammcneilly.pocketleague.ui
+package com.adammcneilly.pocketleague.composables
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.BarChart
@@ -9,9 +9,11 @@ import androidx.compose.material3.NavigationRail
 import androidx.compose.material3.NavigationRailItem
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.adammcneilly.pocketleague.shared.screens.Level1Navigation
 import com.adammcneilly.pocketleague.shared.screens.Navigation
 import com.adammcneilly.pocketleague.shared.screens.ScreenIdentifier
+
 
 /**
  * A navigation rail is similar to a [Level1BottomBar] except it will show up on the left side of the
@@ -19,9 +21,12 @@ import com.adammcneilly.pocketleague.shared.screens.ScreenIdentifier
  */
 @Composable
 fun Navigation.Level1NavigationRail(
-    selectedTab: ScreenIdentifier
+    selectedTab: ScreenIdentifier,
+    modifier: Modifier = Modifier,
 ) {
-    NavigationRail {
+    NavigationRail(
+        modifier = modifier,
+    ) {
         NavigationRailItem(
             icon = { Icon(Icons.Default.Feed, "Feed") },
             label = { Text("FEED") },
